fix(examples): 修复 progress_demo 剩余的 int64 参数类型错误
cliui/progress v0.3.0 的 Txt/DynamicText/WithMaxSteps 参数已改为 int64，
补齐 RoundTrip().WithMaxSteps、Txt、DynamicText 三处的 int64 转换，
使 ./_examples/cliapp 可正常编译。

+++ _examples/cmd/progress_demo.go
@@ -115,7 +115,7 @@ func showAllProgressBar(maxStep int) {
 }
 
 func runRoundTripBar(max int) {
-	p := progress.RoundTrip(0).WithMaxSteps(max)
+	p := progress.RoundTrip(0).WithMaxSteps(int64(max))
 
 	// running
 	runProgressBar(p, max, 120)
@@ -124,7 +124,7 @@ func runRoundTripBar(max int) {
 }
 
 func txtProgressBar(maxStep int) {
-	txt := progress.Txt(maxStep)
+	txt := progress.Txt(int64(maxStep))
 	txt.AddMessage("message", "Handling ... ")
 	// txt.Overwrite = false
 	// running
@@ -144,7 +144,7 @@ func dynamicTextBar(maxStep int) {
 	}
 
 	// maxStep = 10
-	p := progress.DynamicText(messages, maxStep)
+	p := progress.DynamicText(messages, int64(maxStep))
 	// p.Overwrite = false
 
 	// running
